// packages/cantonjs-splice-validator/src/generated/scanProxy.types.ts
export interface components {
    schemas: {
// ...
        AnsEntry: {
            /**
             * @description If present, Daml contract ID of template `Splice.Ans:AnsEntry`.
             *     If absent, this is a DSO-provided entry for either the DSO or an SV.
             */
            contract_id?: string;
            /** @description Owner party ID of this ANS entry. */
            user: string;
            /** @description The ANS entry name. */
            name: string;
            /** @description Either empty, or an http/https URL supplied by the `user`. */
            url: string;
            /** @description Arbitrary description text supplied by `user`; may be empty. */
            description: string;
            /**
             * Format: date-time
             * @description Time after which this ANS entry expires; if renewed, it will have a
             *     new `contract_id` and `expires_at`.
             *     If `null` or absent, does not expire; this is the case only for
             *     special entries provided by the DSO.
             */
            expires_at?: string;
        };
// ...
    };
// ...
}
export type $defs = Record<string, never>;
export interface operations {
    getDsoPartyId: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description ok */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["GetDsoPartyIdResponse"];
                };
            };
        };
    };
    getDsoInfo: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description ok */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["GetDsoInfoResponse"];
                };
            };
        };
    };
// ...
    getOpenAndIssuingMiningRounds: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description ok */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["GetOpenAndIssuingMiningRoundsProxyResponse"];
                };
            };
        };
    };
// ...
    lookupAnsEntryByParty: {
        parameters: {
            query?: never;
            header?: never;
            path: {
                party: string;
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description ok */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["LookupEntryByPartyResponse"];
                };
            };
            404: components["responses"]["404"];
        };
    };
    listAnsEntries: {
        parameters: {
            query: {
                name_prefix?: string;
                page_size: number;
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description ok */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ListEntriesResponse"];
                };
            };
        };
    };
    lookupAnsEntryByName: {
        parameters: {
            query?: never;
            header?: never;
            path: {
                name: string;
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description ok */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["LookupEntryByNameResponse"];
                };
            };
            404: components["responses"]["404"];
        };
    };
// ...
    getHoldingsSummaryAt: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody: {
            content: {
                "application/json": components["schemas"]["HoldingsSummaryRequest"];
            };
        };
        responses: {
            /** @description ok */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["HoldingsSummaryResponse"];
                };
            };
            400: components["responses"]["400"];
            404: components["responses"]["404"];
            500: components["responses"]["500"];
        };
    };
    listUnclaimedDevelopmentFundCoupons: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description ok */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ListUnclaimedDevelopmentFundCouponsResponse"];
                };
            };
        };
    };
}